+++ ReleaseNotes/release_notes.md
@@ -1,4 +1,6 @@
 ## New features:
+  1. add `Ostap::Kinematics::Dalitz0::E1,E2,E3` and `Ostap::Kinematics::Dalitz0::P1,P2,P3` three-argument methods, They are less efficient than corresponding two-argument methods of class `Ostap:Kinematics::Dalitz`
+
 
 ## Backward incompatible